api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv==1.1.0
pydantic==2.11
sqlalchemy==2.0.40
asyncpg==0.30.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-asyncio==0.26.0
httpx==0.28.1
pytest-cov==6.1.1
aiosqlite==0.21.0
